:art: [#14] fixed broken test_file naming
