+++ src/woo_publications/api/drf_spectacular/extensions.py
@@ -6,17 +6,15 @@ from woo_publications.api.utils import underscore_to_camel
 class CamelizeFilterExtension(DjangoFilterExtension):
     priority = 1
 
-    def get_schema_operation_parameters(self, auto_schema, *args, **kwargs):
+    def get_schema_operation_parameters(
+        self, auto_schema, *args, **kwargs
+    ):  # pragma: no cover
         """
         camelize query parameters
         """
-        parameters = super().get_schema_operation_parameters(  # pragma: no cover
+        parameters = super().get_schema_operation_parameters(
             auto_schema, *args, **kwargs
         )
-
         for parameter in parameters:
-            parameter["name"] = underscore_to_camel(
-                parameter["name"]
-            )  # pragma: no cover
-
-        return parameters  # pragma: no cover
+            parameter["name"] = underscore_to_camel(parameter["name"])
+        return parameters
