+++ src/woo_publications/api/urls.py
@@ -4,13 +4,13 @@ from django.views.generic import RedirectView
 from drf_spectacular.views import SpectacularJSONAPIView, SpectacularRedocView
 from rest_framework import routers
 
-from woo_publications.metadata.api.viewsets import InformationCategoryViewset
+from woo_publications.metadata.api.viewsets import InformationCategoryViewSet
 
 app_name = "api"
 
 router = routers.DefaultRouter()
 
-router.register("informatie-categorie", InformationCategoryViewset)
+router.register("informatiecategorieen", InformationCategoryViewSet)
 
 urlpatterns = [
     path("docs/", RedirectView.as_view(pattern_name="api:api-docs")),
